Demonstrate basic monorepo layout
This POC shows what I am hoping to get out of this repo: a lightweight
way to produce small NPM modules with some common infrastructure for
dev-time concerns like code formatting and testing. These two modules
are extracted out of the "dented" package, which I will deprecate in
favor of these two scoped packages. Build not set-up yet, but this is
enough to demonstrate the idea.