Run flow against __tests_ as well
Some notes about this:

    # Naive, doesn't work because of loose flow-bin@* requirement:
    npx flow-typed install jest@24.0.0-alpha.12

    # Doesn't work because latest def is for v23:
    npx flow-typed install --flowVersion 0.91.0 jest@24.0.0-alpha.12

    # Doesn't work because of loose flow-bin@* requirement:
    npx flow-typed create-stub jest@24.0.0-alpha.12

    # Doesn't work because create-stub doesn't support --flowVersion:
    npx flow-typed create-stub --flowVersion 0.91.0 jest@24.0.0-alpha.12

    # Works but creates basically useless stub:
    yarn add -W flow-bin@0.91.0
    npx flow-typed create-stub jest@24.0.0-alpha.12

    # Pulls down most recent, most useful libdef that basically works:
    npx flow-typed install jest@23
    yarn add -W flow-bin@*

One sad thing to note though, this libdef defines the Jest globals (like
`describe`, `it` etc) and Flow is happy, but because Flow doesn't allow
us to scope config to specific directories, it means that it won't catch
errors like referring to `it` and so on outside of the test suite.
Still, better than not having checks at all. Related Flow issue is, I
think, this one:

    https://github.com/facebook/flow/issues/4738