Add delay package
This starts to get a bit messy. First use of async means that we have to
dance around a bit to avoid the use of the regeneratorRuntime. Will need
to solve the problem in the built packages differently (probably
including the polyfill or something) when we start adding async
functions.

Also note the hackery to get fake timers behaving reasonably with
promises.