Add a script that warns about undeclared dependencies
Because it is a bit of a foot gun that we can add dependencies at the
root and "see" them in the dev environment, but in the published
packages fail because of undeclared dependencies.

Note that we scan "lib/" and not "src/" because Babel may add requires
for packages like "@babel/runtime" that won't be visible anywhere under
"src/" but are very much needed at runtime.

As you can see in this commit, it already caught an issue for us, which
we fix at the same time. Sample output from the run:

```
$ scripts/check-dependencies
Scanning for undeclared package dependencies:

  debounce: OK
  dedent: OK
  delay: OK
  escapeHTML: OK
  indent: OK
  isObject: OK
  nullthrows: OK
  stableStringify: MISSING
    @babel/runtime-corejs2

Add missing dependencies with:

(cd packages/stableStringify && yarn add @babel/runtime-corejs2)

zsh: exit 1     scripts/check-dependencies
```