Switch from Flow to TypeScript
Not saying this is going to be a permanent switch (who am I kidding?);
but I have to give a presentation on JS type systems next week, and
while I have years of experience with Flow, my experimentation with
TypeScript has been superficial. So moving stuff over in this project to
see how it all works/feels.

Frankly, pretty amazed at how everything works (or seems to) with
relatively straightforward changes. Tests, lints, typechecking (duh),
building, formatting - it all seems to work.

Points of note in what is otherwise a set of obvious changes:

- ESLint's no-unused-vars rule was tripping up on TypeScript's fake
  "this" parameter type annotation, so I had to add it to the ignore
  pattern.
- Makefile needed tweaking because we can't pass the ".mjs" extension to
  the TypeScript compiler; all the files needed to be renamed and our
  `--keep-file-extension` trick needs to be augmented with a `mv` after
  the build to rename the compiled ".ts" to ".mjs".
- TypeScript cast syntax is much nicer, the utility types are pretty and
  useful (eg. `ReturnType<>`, `NonNullable<>`), and DefinitelyTyped
  (that's where I got `@types/jest` from) is amazingly useful. Many of
  the files barely needed touching. Type guards seem to work really
  well: for example, I was able to remove a `$FlowIssue` that was needed
  with my old `isObject()` implementation (maybe I could have made that
  better with Flow `%checks` predicate, but I don't think I was able to
  make it work).
- I will miss `@flow strict`.