build: don't copy .d.ts files to .mjs
The build looks like this:

1. Make lib/index.d.ts from src/index.ts.
2. Make lib/index.js from src/index.ts.
3. Make lib/index.ts from src/index.ts.
4. Rename lib/index.ts to lib/index.mjs.
5. Copy src/index.js.flow to lib/index.js.flow.

Step 4 was buggy, creating an unwanted lib/index.d.mjs as well.

Note that the confusing build steps are necessitated because we want to
include a "main" output (".js"), an ES modules outpus (".mjs"), a Flow
definitions files (".js.flow") and a TypeScript one too (".d.ts"), but
you can't control what file extension Babel will emit (it will set it
to ".js", or keep the original, but you can't specify an arbitrary
extension, so we need the rename).