Make all packages delegate to workspace-scripts
Except workspace-scripts itself, of course, which just runs its own
"bin/index.js".

Note there is a chicken-and-egg thing here, which is that until you
"build" none of this will probably work, and that's why I may end up
keeping the build as a Makefile step at the top level rather than
pulling it in to workspace-scripts (although I might end up using nested
Makefiles so that you can build an individual package if you want).