Add invariant transform
Not using it yet though because we have a bit of a chicken and egg
problem. I can't add it to the default list of plug-ins because it might
not be built yet. Originally it was in vanilla JS but I ported it to
TypeScript. But I think if I can't come up with a way to bootstrap it in
the Makefile before the other transforms run then I'll just change it
back and do it the real dumb way.