Upgrade packages
`yarn upgrade-interactive` all the things.

Note that because the non-dev dependencies in the individual packages
are all locked to a specific version, I also ran `yarn outdated` to
identify those, and then:

    (cd packages/workspace-scripts && yarn add @babel/parser@7.4.2 @babel/runtime-corejs2@7.4.2 @babel/traverse@7.4.0)
    for X in babel-plugin-invariant-transform dedent delay delay event-emitter stable-stringify; do
      (cd packages/$X && yarn add @babel/runtime-corejs2@7.4.2)
    done

will evidently want to make a less manual way to do that via a script.

Includes tweaks in workspace-scripts/src/publish.ts to keep the new
version of TypeScript happy.